requirements.txt
accelerate~=0.30.0
bitsandbytes~=0.43.1
huggingface-hub~=0.23.0
langchain~=0.1.17
langchain~=0.1.17
langchain-community~=0.0.36
langchain-core~=0.1.50
langchain-openai~=0.1.6
langchain-text-splitters~=0.0.1
langchain-weaviate~=0.0.1.post1
langchainhub~=0.1.15
langsmith~=0.1.53
pypdf~= 4.2.0
python-dotenv~=1.0.1
sentence-transformers~=2.7.0
streamlit~=1.34.0
tensorrt~=10.0.1
tiktoken~=0.6.0
tokenizers~=0.19.1
torch~=2.3.0 --index-url https://download.pytorch.org/whl/cu118
torch~=2.3.0+cu118
torchaudio~=2.3.0 --index-url https://download.pytorch.org/whl/cu118
torchvision~=0.18.0 --index-url https://download.pytorch.org/whl/cu118
transformers~=4.40.1
weaviate-client~=4.5.7
